refactor: extract openaiProvider to provider_openai.go
